rename config -> dump_config
