+++ lib/llm/src/http/service.rs
@@ -25,7 +25,7 @@ pub mod error;
 pub mod health;
 pub mod metrics;
 pub mod service_v2;
-pub mod config;
+pub mod dump_config;
 
 pub use axum;
 pub use metrics::Metrics;
